refactor : 불필요한 주석 및 사용하지 않는 스타일 규칙 제거, 스타일 규칙을 해당 페이지의 css파일로 이동
