+++ src/components/card/CardSearchBar.css
@@ -1,0 +1,20 @@
+.search-bar{
+    display: flex;
+    background-color: #F5F5F5;
+    padding: 1.5rem;
+    margin: 0 3.2rem;
+    border-radius: 10px;
+    max-width: 113rem;
+    width: 90%;
+    gap: 1rem;
+}
+
+.search-input{
+    border: 0;
+    background-color: #F5F5F5;
+    width: 100%;
+}
+
+.search-input:focus{
+    outline:none;
+}
