+++ src/components/footer/FooterStyle.css
@@ -39,7 +39,6 @@ footer span{
     color: #fff;
 }
 
-
 /* 반응형 */
 /* Mobile */
 @media (max-width: 768px){
