+++ src/components/landingSection/LandingSection.css
@@ -1,4 +1,3 @@
-/* section */
 section {
     display: grid;
     padding: 3.12rem 10rem;
@@ -55,12 +54,10 @@ section p{
     grid-area: description;
 }
 
-
 section img{
     width: 55rem;
     grid-area: image;
 }
-
 
 /* section span 그라데이션 배경색 */
 .save span{
@@ -85,7 +82,6 @@ section .font-gradation,
     -webkit-background-clip: text;
     -webkit-text-fill-color: transparent;
 }
-
 
 /* 반응형 */
 /* Tablet */
