+++ src/pages/login/LoginPage.css
@@ -12,7 +12,6 @@ html{
     background-color: #F0F6FF;
 }
 
-/* form */
 label{
     display: block;
     font-size: 1.4rem;
