+++ src/pages/shared/SharedPage.css
@@ -1,15 +1,3 @@
-
-/* main */
-.sharedMain{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    background-color: #fff;
-    padding: 4rem 0;
-}
-
-/* folder 정보 */
 .folder-info{
     padding: 9.6rem  0 6rem;
     display: flex;
@@ -42,48 +30,12 @@
     font-weight: 600;
 }
 
-/* folder 검색 */
-.search-bar{
+.sharedMain{
     display: flex;
-    background-color: #F5F5F5;
-    padding: 1.5rem;
-    margin: 0 3.2rem;
-    border-radius: 10px;
-    max-width: 113rem;
-    width: 90%;
-    gap: 1rem;
-}
-
-.search-input{
-    border: 0;
-    background-color: #F5F5F5;
-    width: 100%;
-}
-
-.search-input:focus{
-    outline:none;
-}
-
-.card-container{
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: center;
-    max-width: 113rem;
-    gap: 2.5rem 2rem;
-    padding: 4rem 3.2rem 0;
-}
-
-/* login User 정보*/
-.user-profile{
-    display: flex;
+    flex-direction: column;
     justify-content: center;
     align-items: center;
-    gap: .6rem;
+    background-color: #fff;
+    padding: 4rem 0;
 }
 
-.user-profile img{
-    width: 2.8rem;
-    height: 2.8rem;
-    border-radius: 2rem;
-}
